HSEARCH-3123 Elasticsearch - Move multitenancy strategies to a separate package
Multitenancy strategies are used in many places, and the .impl package
already has lots of dependencies (since it hosts the backend).
If we want to avoid an overly complex package dependency graph, we must
move multitenancy to its own package.
