+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/document/impl/ElasticsearchDocumentObjectBuilder.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 import org.hibernate.search.v6poc.backend.document.DocumentElement;
 import org.hibernate.search.v6poc.backend.elasticsearch.document.model.impl.ElasticsearchIndexSchemaObjectNode;
 import org.hibernate.search.v6poc.backend.elasticsearch.gson.impl.JsonAccessor;
-import org.hibernate.search.v6poc.backend.elasticsearch.impl.MultiTenancyStrategy;
+import org.hibernate.search.v6poc.backend.elasticsearch.multitenancy.impl.MultiTenancyStrategy;
 import org.hibernate.search.v6poc.logging.impl.Log;
 import org.hibernate.search.v6poc.util.impl.common.LoggerFactory;
