+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/impl/ElasticsearchBackendImpl.java
@@ -16,6 +16,7 @@ import org.hibernate.search.v6poc.backend.elasticsearch.document.impl.Elasticsea
 import org.hibernate.search.v6poc.backend.elasticsearch.document.model.impl.ElasticsearchRootIndexSchemaCollectorImpl;
 import org.hibernate.search.v6poc.backend.elasticsearch.index.impl.ElasticsearchIndexManagerBuilder;
 import org.hibernate.search.v6poc.backend.elasticsearch.index.impl.IndexingBackendContext;
+import org.hibernate.search.v6poc.backend.elasticsearch.multitenancy.impl.MultiTenancyStrategy;
 import org.hibernate.search.v6poc.backend.elasticsearch.search.query.impl.SearchBackendContext;
 import org.hibernate.search.v6poc.backend.elasticsearch.logging.impl.Log;
 import org.hibernate.search.v6poc.backend.elasticsearch.orchestration.impl.ElasticsearchWorkOrchestrator;
